practice_2/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2024.10
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==5.1.6
django-cache-url==3.4.5
django-ckeditor-5==0.2.17
django-crispy-forms==2.3
django-filter==25.1
django-js-asset==3.1.2
environs==14.1.1
marshmallow==3.26.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
python-dotenv==1.0.1
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
